change type item quantity to int 64

+++ be/entity/items.go
@@ -11,11 +11,11 @@ type Item struct {
 	ID          uuid.UUID `json:"id" gorm:"type:uuid;primaryKey"`
 	Name        string    `json:"name" gorm:"unique;not null"`
 	Description string    `json:"description" gorm:"type:text"`
-	Quantity    int       `json:"quantity" gorm:"default:0"`
+	Quantity    int64     `json:"quantity" gorm:"default:0"`
 	Price       float64   `json:"price" gorm:"default:0"`
 	Picture     string    `json:"picture"`
 	View        int64     `json:"view" gorm:"default:0"`
-	Recommend   int       `json:"recommend" gorm:"default:0"`
+	Recommend   int64     `json:"recommend" gorm:"default:0"`
 	CreatedAt   time.Time `json:"created_at"`
 	UpdatedAt   time.Time `json:"updated_at"`
 }
